+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+- Save and restore selected state of keyframes.
+
 ## [0.12.0] - 2021-10-19
 ### Changed
 - Save and restore selected state of some entities.